ALSA: scarlett2: Fix speaker switching and mute notifications
When speaker switching is enabled, the values of the line 1-4 output
controls can change. The notification was sent with a mask of only
SNDRV_CTL_EVENT_MASK_INFO. Change the notification to set the
SNDRV_CTL_EVENT_MASK_VALUE mask bit as well.

When the mute control is updated, the notification was sent with a
mask of SNDRV_CTL_EVENT_MASK_INFO. Change the mask to the correct
value of SNDRV_CTL_EVENT_MASK_VALUE.

Signed-off-by: Geoffrey D. Bennett <g@b4.vu>